requirements.txt
pydantic
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-multipart
pdfplumber
sentence-transformers
numpy
qdrant-client
llama-cpp-python
huggingface_hub
json_repair
